infinite_grid: fix: depth not calculated properly
Earlier was calculating the clip space depth only, not the frag depth
that is required. This is done so by using the inbuilt gl_DepthRange
uniform that provides far and near, these are not the same as the
camera far and near planes. A simple calculation later, we get the
correct depth :)