fix: examples: opengl context not set so shader compile error
The program fails to link and gives an error that the vertex shader
doesn't have a main(). This is an error that is caused because the
context is not set correctly but the error is not specific.  For more
info, see https://bugs.freedesktop.org/show_bug.cgi?id=110872

One way to prevent shader caching is to export
MESA_GLSL_CACHE_DISABLE=1 before running the program, and now it
throws an error that glsl 330 is not support making it clear that it
is context related bug. So just set the context and the error
disappears fixing the bug.