bvh: BVHTree: ray_cast(): change callback requirements
This is a breaking change but it is necessary for it to be extensible.

The callback now requires `FnMut(T) -> Option<RayHitData<T,
ExtraData>> + std::marker::Copy`. This releases the user from messing
with lifetimes but leads to the user having to access the ray from
outside. It seems like a worthwhile trade-off. The callback function
is no longer taken by reference but requires std::marker::Copy be
implemented which should be the case for basically all closures and
functions, thus making it more ergonomic.