app: App: init(): take generic init data from outside the environment
This allows the user to pass any data they want to [`App::init()`].

This will be useful when the GPU context needs to be made outside the
App but passed to the App.

TODO: Environment should allow shared & mutable reference to the
environment whenever needed. This way the init data passed to init()
can also have some initialization that requires the environment to be
initialized.

TODO: App::update() should return if the app should close and return
any data in the process.