app: Environment: run(): take self as mutable reference
This allows the [`App`] to be dropped before the environment is
dropped. Useful for any OpenGL application, prior to this change, the
context would be dropped when the window is dropped which happens when
the environment is dropped. Since the function returns the [`App`],
the [`Environment`] is dropped before the [`App`] leading to any
objects that require the OpenGL context to not function
properly (generally a segfault).